fix profile and auth
