+++ testsbase/test.py
@@ -413,257 +413,3 @@ class TestPremieresPage(BasePage):
             raise Exception("title does not equal", title, self.PREMIERES_TITLE)
 
         assert self.find((By.CLASS_NAME, self.PREMIERES_DAY_CLASS_NAME), 3)
-
-
-class TestSignup(BasePage):
-    LOGIN = 'Qwe123@a.a'
-    PASSWD = 'Qwe123@a.a'
-    X_BUTTON_OPEN_REG_WINDOW = '/html/body/div/div[1]/div/div/div/div[2]/div/a'
-    X_BUTTON_OPEN_LOGIN_WINDOW = '/html/body/div/div[1]/div/div/div/div[1]/div/a'
-    X_INPUT_NICKNAME = '/html/body/div/div[1]/div/div/div/div[2]/form/div[1]/input'
-    X_INPUT_EMAIL = '/html/body/div/div[1]/div/div/div/div[2]/form/div[2]/input'
-    X_INPUT_PASSWORD = '/html/body/div/div[1]/div/div/div/div[2]/form/div[3]/input'
-    X_INPUT_REPEAT_PASSWORD = '/html/body/div/div[1]/div/div/div/div[2]/form/div[4]/input'
-    X_BUTTON_REG = '/html/body/div/div[1]/div/div/div/div[2]/form/button'
-    X_ALREADY_REGISTER = '/html/body/div/div[1]/div/div/div/div[2]/form/div[2]/div'
-    CLASS_WRONG_INPUT = 'modal__input_red_border'
-    NICKNAME = 'Admin'
-    URL_PAGE_LOGIN = f'{DOMAIN}/login/'
-    URL_PAGE_SIGNUP = f'{DOMAIN}/signup/'
-
-    def test_redirect_login(self):
-        self.render(self.URL_PAGE_SIGNUP)
-
-        btn_login = self.find((By.XPATH, self.X_BUTTON_OPEN_LOGIN_WINDOW), 10)
-        btn_login.click()
-        pattern = self.URL_PAGE_LOGIN
-        current__url = dvr.get_instance().current_url
-        if (pattern != current__url):
-            raise Exception("wrong redirect", pattern, current__url)
-
-    def test_short_nick(self):
-        self.render(self.URL_PAGE_SIGNUP)
-        short_nick = '123'
-
-        input_nickname = self.find((By.XPATH, self.X_INPUT_NICKNAME))
-        input_nickname.send_keys(short_nick)
-        input_email = self.find((By.XPATH, self.X_INPUT_EMAIL)).send_keys(self.LOGIN)
-        input_password = self.find((By.XPATH, self.X_INPUT_PASSWORD)).send_keys(self.PASSWD)
-        input_repeat_password = self.find((By.XPATH, self.X_INPUT_REPEAT_PASSWORD)).send_keys(self.PASSWD)
-
-        self.find((By.XPATH, self.X_BUTTON_REG)).click()
-
-        is_wrong = self.CLASS_WRONG_INPUT in input_nickname.get_attribute("class").split()
-        if (is_wrong != True):
-            raise Exception("wrong check", input_nickname, short_nick)
-
-    def test_invalid_emeil(self):
-        self.render(self.URL_PAGE_SIGNUP)
-        wrong_email = '123456'
-
-        input_nickname = self.find((By.XPATH, self.X_INPUT_NICKNAME)).send_keys(self.NICKNAME)
-        input_email = self.find((By.XPATH, self.X_INPUT_EMAIL))
-        input_email.send_keys(wrong_email)
-        input_password = self.find((By.XPATH, self.X_INPUT_PASSWORD)).send_keys(self.PASSWD)
-        input_repeat_password = self.find((By.XPATH, self.X_INPUT_REPEAT_PASSWORD)).send_keys(self.PASSWD)
-
-        self.find((By.XPATH, self.X_BUTTON_REG)).click()
-
-        is_wrong = self.CLASS_WRONG_INPUT in input_email.get_attribute("class").split()
-        if (is_wrong != True):
-            raise Exception("wrong check", input_email, wrong_email)
-
-    def test_invalid_password(self):
-        self.render(self.URL_PAGE_SIGNUP)
-        wrong_password = 'abcds'
-
-        input_nickname = self.find((By.XPATH, self.X_INPUT_NICKNAME)).send_keys(self.NICKNAME)
-        input_email = self.find((By.XPATH, self.X_INPUT_EMAIL)).send_keys(self.LOGIN)
-        input_password = self.find((By.XPATH, self.X_INPUT_PASSWORD))
-        input_password.send_keys(wrong_password)
-        input_repeat_password = self.find((By.XPATH, self.X_INPUT_REPEAT_PASSWORD))
-        input_repeat_password.send_keys(wrong_password)
-
-        self.find((By.XPATH, self.X_BUTTON_REG)).click()
-
-        is_wrong_input_password = self.CLASS_WRONG_INPUT in input_password.get_attribute("class").split()
-        is_wrong_input_repeat_password = self.CLASS_WRONG_INPUT in input_password.get_attribute("class").split()
-        if (is_wrong_input_password != True or is_wrong_input_repeat_password != True):
-            raise Exception("wrong check", input_password, wrong_password)
-
-    def test_correct_register(self):
-        self.render(self.URL_PAGE_SIGNUP)
-
-        input_nickname = self.find((By.XPATH, self.X_INPUT_NICKNAME)).send_keys(self.NICKNAME)
-        input_email = self.find((By.XPATH, self.X_INPUT_EMAIL)).send_keys(self.LOGIN)
-        input_password = self.find((By.XPATH, self.X_INPUT_PASSWORD)).send_keys(self.PASSWD)
-        input_repeat_password = self.find((By.XPATH, self.X_INPUT_REPEAT_PASSWORD)).send_keys(self.PASSWD)
-
-        self.find((By.XPATH, self.X_BUTTON_REG)).click()
-        msg_already_register = self.find((By.XPATH, self.X_ALREADY_REGISTER)).text
-        if (msg_already_register != "Пользователь с таким email уже зарегистрирован"):
-            raise Exception("wrong register msg", msg_already_register,
-                            "Пользователь с таким email уже зарегистрирован")
-
-
-class TestLogin(BasePage):
-    LOGIN = 'Qwe123@a.a'
-    PASSWD = 'Qwe123@a.a'
-    X_BUTTON_OPEN_REG_WINDOW = '/html/body/div/div[1]/div/div/div/div[2]/div/a'
-    X_BUTTON_OPEN_LOGIN_WINDOW = '/html/body/div/div[1]/div/div/div/div[1]/div/a'
-    X_INPUT_EMAIL = '/html/body/div/div[1]/div/div/div/div[1]/form/div[1]/input'
-    X_INPUT_PASSWORD = '/html/body/div/div[1]/div/div/div/div[1]/form/div[2]/input'
-    X_BUTTON_LOGIN = '/html/body/div/div[1]/div/div/div/div[1]/form/button'
-    CLASS_WRONG_INPUT = 'modal__input_red_border'
-    CLASS_MODAL = 'modal__background'
-    URL_PAGE_LOGIN = f'{DOMAIN}/login/'
-    URL_PAGE_SIGNUP = f'{DOMAIN}/signup/'
-
-    def test_redirect_signup(self):
-        self.render(self.URL_PAGE_LOGIN)
-
-        btn_login = self.find((By.XPATH, self.X_BUTTON_OPEN_REG_WINDOW), 10)
-        btn_login.click()
-        pattern = self.URL_PAGE_SIGNUP
-        current__url = dvr.get_instance().current_url
-        if (pattern != current__url):
-            raise Exception("wrong redirect", pattern, current__url)
-
-    def test_invalid_emeil(self):
-        self.render(self.URL_PAGE_LOGIN)
-        wrong_email = '123456'
-
-        input_email = self.find((By.XPATH, self.X_INPUT_EMAIL))
-        input_email.send_keys(wrong_email)
-        input_password = self.find((By.XPATH, self.X_INPUT_PASSWORD)).send_keys(self.PASSWD)
-
-        self.find((By.XPATH, self.X_BUTTON_LOGIN)).click()
-
-        is_wrong = self.CLASS_WRONG_INPUT in input_email.get_attribute("class").split()
-        if (is_wrong != True):
-            raise Exception("wrong check", input_email, wrong_email)
-
-    def test_invalid_password(self):
-        self.render(self.URL_PAGE_LOGIN)
-        wrong_password = '123456'
-
-        input_email = self.find((By.XPATH, self.X_INPUT_EMAIL)).send_keys(self.LOGIN)
-        input_password = self.find((By.XPATH, self.X_INPUT_PASSWORD))
-        input_password.send_keys(wrong_password)
-
-        self.find((By.XPATH, self.X_BUTTON_LOGIN)).click()
-
-        is_wrong = self.CLASS_WRONG_INPUT in input_password.get_attribute("class").split()
-        if (is_wrong != True):
-            raise Exception("wrong check", input_password, input_password)
-
-    def test_correct_login(self):
-        self.render(self.URL_PAGE_LOGIN)
-
-        self.find((By.XPATH, self.X_INPUT_EMAIL)).send_keys(self.LOGIN)
-        self.find((By.XPATH, self.X_INPUT_PASSWORD)).send_keys(self.PASSWD)
-
-        self.find((By.XPATH, self.X_BUTTON_LOGIN)).click()
-        try:
-            self.find((By.CLASS, self.CLASS_MODAL))
-        except:
-            None
-
-
-class TestProfile(BasePage):
-    CLASS_CHANGE_SVG = 'profile__change__svg'
-    CLASS_FORM_CHANGE = 'profile__change'
-    CLASS_FORM_DISPLAY_NONE = 'dysplay-none'
-    CLASS_FORM_DISPLAY_FLEX = 'dysplay-flex'
-    X_SAVE_NEW_NAME = '/html/body/div/div/div[2]/div[2]/div[2]/form[1]/button'
-    URL_PAGE_PROFILE = f'{DOMAIN}/profile/'
-
-    @needed_auth
-    def test_open_change_input(self):
-
-        self.render(self.URL_PAGE_PROFILE)
-        self.find((By.CLASS_NAME, self.CLASS_CHANGE_SVG)).click()
-        form_change = self.find((By.CLASS_NAME, self.CLASS_FORM_CHANGE))
-        is_opened = self.CLASS_FORM_DISPLAY_FLEX in form_change.get_attribute("class").split()
-        if (is_opened != True):
-            raise Exception("can't open form change", form_change)
-        self.find((By.CLASS_NAME, self.CLASS_CHANGE_SVG)).click()
-        is_closed = self.CLASS_FORM_DISPLAY_NONE in form_change.get_attribute("class").split()
-        if (is_closed != True):
-            raise Exception("can't closen form change", form_change)
-
-    @needed_auth
-    def test_change_name(self):
-
-        self.render(self.URL_PAGE_PROFILE)
-        self.find((By.CLASS_NAME, self.CLASS_CHANGE_SVG), 3).click()
-        field_Input_new_name = self.find((By.XPATH, "//input[@placeholder='Введите новое имя пользователя']"), 3)
-        field_Input_new_name.send_keys('Zxc543')
-        self.find((By.XPATH, self.X_SAVE_NEW_NAME), 4).click()
-
-        new_name = self.find((By.XPATH, "//div[contains(text(), 'Zxc543')]"), 3)
-
-        if (bool(new_name) != True):
-            raise Exception("can't change name", new_name)
-
-        self.render(self.URL_PAGE_PROFILE)
-        self.find((By.CLASS_NAME, self.CLASS_CHANGE_SVG), 3).click()
-        field_new_name = self.find((By.XPATH, "//input[@placeholder='Введите новое имя пользователя']"), 3)
-        field_new_name.send_keys('Admin')
-        self.find((By.XPATH, self.X_SAVE_NEW_NAME), 3).click()
-
-        new_name = self.find((By.XPATH, "//div[contains(text(), 'Admin')]"), 3)
-        if (bool(new_name) != True):
-            raise Exception("can't return name", new_name)
-
-    @needed_auth
-    def test_check_value_num_of_rates(self):
-
-        self.render(self.URL_PAGE_PROFILE)
-
-        field_num_of_rates = self.find((By.XPATH, "//div[contains(text(), 'Оценок:')]//following-sibling::div"), 3).text
-        if (not field_num_of_rates):
-            raise Exception("empty field")
-        if (not (field_num_of_rates == "нет оценок" or int(field_num_of_rates) > -1)):
-            raise Exception("wrong format")
-
-    @needed_auth
-    def test_check_value_num_of_coll(self):
-
-        self.render(self.URL_PAGE_PROFILE)
-
-        field_num_of_coll = self.find((By.XPATH, "//div[contains(text(), 'Коллекций:')]//following-sibling::div"),
-                                      3).text
-        if (not field_num_of_coll):
-            raise Exception("empty field")
-        if (not (int(field_num_of_coll) > -1)):
-            raise Exception("wrong format")
-
-    @needed_auth
-    def test_check_value_num_of_rewiews(self):
-
-        self.render(self.URL_PAGE_PROFILE)
-
-        field_num_of_rewiews = self.find((By.XPATH, "//div[contains(text(), 'Рецензий:')]//following-sibling::div"),
-                                         3).text
-        if (not field_num_of_rewiews):
-            raise Exception("empty field")
-        if (not (field_num_of_rewiews == "нет рецензий" or int(field_num_of_rewiews) > -1)):
-            raise Exception("wrong format")
-
-    @needed_auth
-    def test_check_main_fields(self):
-
-        self.render(self.URL_PAGE_PROFILE)
-        date_of_reg = self.find((By.XPATH, "//div[contains(text(), 'Дата регистрации:')]"), 5)
-        if not bool(date_of_reg):
-            raise Exception("can't find 'Дата регистрации:' field")
-        num_of_rates = self.find((By.XPATH, "//div[contains(text(), 'Оценок:')]"), 5)
-        if not bool(num_of_rates):
-            raise Exception("can't find 'Оценок:' field")
-        num_of_coll = self.find((By.XPATH, "//div[contains(text(), 'Коллекций:')]"), 5)
-        if not bool(num_of_coll):
-            raise Exception("can't find 'Коллекций:' field")
-        num_of_rewiews = self.find((By.XPATH, "//div[contains(text(), 'Рецензий:')]"), 5)
-        if not bool(num_of_rewiews):
-            raise Exception("can't find 'Рецензий:' field")
